requirements.txt
fastapi==0.115.0
uvicorn==0.30.1
python-multipart==0.0.9

# RAG + embeddings
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
numpy==1.26.4
pandas==2.2.2

# TF-IDF fallback
scikit-learn==1.5.0

# Translation + language detection
deep-translator==1.11.4
langdetect==1.0.9

# PDF & document extraction
pdfplumber==0.10.3
pypdf==4.3.1

# NLI + stance
torch
transformers==4.41.2
nltk==3.8.1

# UI
streamlit==1.36.0

# Utilities
tqdm==4.66.4
protobuf==4.25.3
sentencepiece==0.2.0
